+++ tests/testthat/test-reorder-within.R
@@ -1,3 +1,4 @@
+skip_if_not_installed("ggplot2")
 suppressPackageStartupMessages(library(ggplot2))
 
 test_that("Can reorder within", {