+++ NEWS.md
@@ -1,5 +1,7 @@
 # tidytext (development version)
 
+* hunspell is now a suggested dependency, thanks to @MichaelChirico (#221)
+
 # tidytext 0.3.4
 
 * Updated the tidy method for a quanteda `dfm` because of the upcoming release of Matrix (#218)