+++ NEWS.md
@@ -1,6 +1,7 @@
 # tidytext (development version)
 
 * hunspell is now a suggested dependency, thanks to @MichaelChirico (#221)
+* Added `stm()` tidiers for high FREX and lift words (#223)
 
 # tidytext 0.3.4